+++ src/hex_map.rs
@@ -1,4 +1,4 @@
-use crate::Hex;
+use crate::{Hex, HexBounds};
 
 /// Hexagon shaped map with [wraparound] utils.
 ///
@@ -6,10 +6,7 @@ use crate::Hex;
 #[derive(Debug, Clone)]
 #[cfg_attr(feature = "ser_de", derive(serde::Serialize, serde::Deserialize))]
 pub struct HexMap {
-    /// The map radius
-    radius: u32,
-    /// The map center
-    center: Hex,
+    bounds: HexBounds,
     /// The 6 mirror centers, used to wrap coordinates
     mirrors: [Hex; 6],
 }
@@ -20,8 +17,7 @@ impl HexMap {
     /// Creates a new hexagonal map of the given `radius` with [`Hex::ZERO`] as its center
     pub const fn new(radius: u32) -> Self {
         Self {
-            radius,
-            center: Hex::ZERO,
+            bounds: HexBounds::new(Hex::ZERO, radius),
             mirrors: Hex::wraparound_mirrors(radius),
         }
     }
@@ -38,8 +34,10 @@ impl HexMap {
     /// ```
     pub fn with_center(self, center: Hex) -> Self {
         Self {
-            radius: self.radius,
-            center,
+            bounds: HexBounds {
+                center,
+                ..self.bounds
+            },
             mirrors: self.mirrors.map(|h| h + center),
         }
     }
@@ -48,14 +46,14 @@ impl HexMap {
     #[must_use]
     /// Returns the map center coordinates
     pub const fn center(&self) -> Hex {
-        self.center
+        self.bounds.center
     }
 
     #[inline]
     #[must_use]
     /// Returns the map radius
     pub const fn radius(&self) -> u32 {
-        self.radius
+        self.bounds.radius
     }
 
     /// Wraps `hex` in the given map radius.
@@ -65,9 +63,9 @@ impl HexMap {
     /// [article]: https://www.redblobgames.com/grids/hexagons/#wraparound
     #[must_use]
     pub fn wrapped_hex(&self, hex: Hex) -> Hex {
-        let pos = hex - self.center;
-        let pos = pos.wrap_with(self.radius, &self.mirrors);
-        pos + self.center
+        let pos = hex - self.center();
+        let pos = pos.wrap_with(self.radius(), &self.mirrors);
+        pos + self.center()
     }
 
     /// Computes the neighbors of `hex` wrapped in the map bounds.
@@ -82,12 +80,12 @@ impl HexMap {
     #[inline]
     /// Returns the number of hexagons in the map
     pub const fn hex_count(&self) -> usize {
-        Hex::range_count(self.radius)
+        self.bounds.hex_count()
     }
 
     /// Returns an iterator with all the coordinates in the map bounds
     pub fn all_coords(&self) -> impl Iterator<Item = Hex> {
-        self.center.range(self.radius)
+        self.bounds.all_coords()
     }
 }